+++ Sources/Float3/Float3.swift
@@ -109,7 +109,7 @@ extension Float3
 	// MARK: `subscript`-Getter
 	
 	/// Access individual elements of the vector via subscript.
-	public subscript(index:Float) -> Float {
+	public subscript(index:Int) -> Float {
 		switch index {
 			case 0: return self.x
 			case 1: return self.y
